+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## 0.4.0 (Unreleased)
 
+NEW FEATURES:
+
+  - New trigger type: ```instead_of```.
+
 IMPROVEMENTS:
 
-  - DSL: the `error` statement now stops the action and makes Vagrant fail with an error.
-  - Ensure the `run` statement always returns the command output.
+  - DSL: the ```error``` statement now stops the action and makes Vagrant fail with an error.
+  - Ensure the ```run``` statement always returns the command output.
 
 ## 0.3.0 (April 4, 2014)