+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.4.0 (Unreleased)
+## 0.4.0 (May 20, 2014)
 
 NEW FEATURES: